refactor(lib): refactor `chatHelpers.js`
* Split `chatHelpers` into multiple files

* Rename `transformMessage` -> `normalizeMessage`
